+++ java/src/integrationTest/groovy/uk/gov/hmcts/CaseControllerSpecification.groovy
@@ -3,31 +3,26 @@ package uk.gov.hmcts
 import com.fasterxml.jackson.databind.JsonNode
 import com.fasterxml.jackson.databind.ObjectMapper
 import groovy.json.JsonOutput
-import groovy.json.JsonSlurper
 import org.jooq.DSLContext
 import org.jooq.SQLDialect
 import org.jooq.generated.enums.CaseState
-import org.jooq.generated.enums.ClaimState
+import org.jooq.generated.enums.Event
 import org.jooq.impl.DSL
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.dao.DuplicateKeyException
 import org.springframework.http.HttpStatus
-import org.springframework.security.test.context.support.WithMockUser
 import org.springframework.test.web.servlet.MockMvc
 import org.springframework.test.web.servlet.setup.MockMvcBuilders
 import org.springframework.transaction.annotation.Transactional
 import org.springframework.web.context.WebApplicationContext
 import spock.lang.Specification
 import uk.gov.hmcts.ccf.TransitionContext
-import uk.gov.hmcts.ccf.api.CaseActions
 import uk.gov.hmcts.ccf.api.ApiEventCreation
-import uk.gov.hmcts.ccf.api.UserInfo
-import uk.gov.hmcts.ccf.controller.CaseController
+import uk.gov.hmcts.ccf.api.CaseActions
+import uk.gov.hmcts.ccf.controller.kase.CaseController
 import uk.gov.hmcts.unspec.CaseHandlerImpl
 import uk.gov.hmcts.unspec.dto.AddClaim
-import uk.gov.hmcts.unspec.dto.ConfirmService
-import org.jooq.generated.enums.Event
 import uk.gov.hmcts.unspec.event.CloseCase
 import uk.gov.hmcts.unspec.event.CreateClaim
 import uk.gov.hmcts.unspec.event.SubmitAppeal
