+++ java/src/main/java/uk/gov/hmcts/ccf/config/TestDataGenerator.java
@@ -11,7 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.ccf.api.ApiEventCreation;
-import uk.gov.hmcts.ccf.controller.CaseController;
+import uk.gov.hmcts.ccf.controller.kase.CaseController;
 import uk.gov.hmcts.unspec.dto.AddClaim;
 import uk.gov.hmcts.unspec.dto.Company;
 import uk.gov.hmcts.unspec.dto.Organisation;
