--- java/src/main/java/uk/gov/hmcts/ccf/controller/CaseSearchResult.java
@@ -1,13 +1,0 @@
-package uk.gov.hmcts.ccf.controller;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import org.jooq.generated.enums.CaseState;
-
-@AllArgsConstructor
-@Data
-class CaseSearchResult {
-    private Long caseId;
-    private CaseState state;
-    private Long partyCount;
-}
