+++ java/src/main/java/uk/gov/hmcts/ccf/controller/user/UserController.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.ccf.controller;
+package uk.gov.hmcts.ccf.controller.user;
 
 import io.swagger.v3.oas.annotations.Parameter;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
